feat: implement global exception handling and enhance API response structure

+++ backend/src/main/java/com/studentzone/app/study/dto/response/StudySessionResponseDTO.java
@@ -10,7 +10,7 @@ import lombok.Data;
 public class StudySessionResponseDTO {
     private Long id;
     private Long userId;
-    private Long pointsEarned;
+    private Double pointsEarned;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
 }
