+++ backend/src/main/java/com/studentzone/app/user/dto/response/UserPointsResponseDTO.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @Builder
 public class UserPointsResponseDTO {
     private Long id;
-    private Long totalPoints;
+    private Double totalPoints;
 }
